x_fase1/week02_day04/am/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
pillow
numpy
joblib
scikit-learn==1.1.3